+++ src/entities.jl
@@ -173,6 +173,15 @@ halo(ents, outer_delta, inner_delta=nothing; kwargs...) =
 Return a single polygon, an iterator, or `Vector` of `Polygon`s equivalent to `ent`.
 
 If `ent` is a `StyledEntity`, all styles will be applied before conversion to polygons.
+
+Keyword arguments can be used to control how certain entity types are converted to polygons:
+
+  - Entities with curves discretized by [`DeviceLayout.adapted_grid`](@ref) use `max_recursions, max_change, rand_factor, grid_step`
+    (these mainly come from path segment/style pairs where boundaries are derived from arbitrary functions)
+  - `atol` is the absolute tolerance used for discretizing other curves (default `1.0nm`)
+  - `Δθ` can be provided to render circles and ellipses with an angular step rather than `atol`
+  - `rtol` can be provided to render tolerance-controlled (non-`adapted_grid`) curves with tolerance `max(atol, rtol*local_curvature_radius)`
+  - Additional rendering options can be provided for user-defined conditional rendering with [`OptionalStyle`](@ref)
 """
 function to_polygons end